src/main/java/se/uu/ub/cora/diva/tocorastorage/db/DivaDbToCoraOrganisationSuccessorConverter.java
package se.uu.ub.cora.diva.tocorastorage.db;

import java.util.Map;

import se.uu.ub.cora.bookkeeper.data.DataGroup;

public class DivaDbToCoraOrganisationSuccessorConverter
		extends DivaDbToCoraOrganisationAncestryConverter implements DivaDbToCoraConverter {

	@Override
	public DataGroup fromMap(Map<String, String> dbRow) {
		this.dbRow = dbRow;
		if (mandatoryValuesAreMissing()) {
			throw ConversionException.withMessageAndException(
					"Error converting organisation successor to Cora organisation successor: Map does not "
							+ "contain mandatory values for organisation id and prdecessor id",
					null);
		}
		return null;
	}

}
